Reorganize CLI folder
