+++ cli/lib/shell.js
